Improve listener configuration.
We now default the listener when starting the server if one hasn't
been specified. This lets the command line handling work correctly
since the default isn't being configured when the server is created.
Also, the XML config file parsing doesn't need to reset the server
listeners to clear out the default any longer. (And if you don't
specify a listener in the config file, it will use the default now.)