Fix multipart/form-data, rewrite header parsing
Each multipart/form-data part has a set of RFC822 headers separated from the
body of the part by a blank line, and is a good excuse to reuse the main header
parsing code. Doing so exposed bugs around exactly how CRLF was handled. This
includes a fairly comprehensive rewrite of the header parsing that also
simplifies the code and adds tests.

It introduces a more efficient use of buffers. There is now a fixed-size header
buffer, based on a quick survey of what other web servers do. (8K seems common
but should be made configurable.) The idea is that this buffer should be
plumbed through the remaining code so that a single one can be reused for each
request, one buffer per worker thread. That is future work.

The HTTP test suite is in terrible shape and hangs for many tests, making it
essentially useless. I tested this mainly via http-common-test-suite and by
running web-playground.